+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+3.6.0 (xxxx-xx-xx)
+--
+Improvements:
+
+* Core: introduction of the Filesystem component.
+* Core: introduction of the Finder component.
+
 3.5.2 (xxxx-xx-xx)
 --
 Improvements: